Update wormhole adaptor (#727)
* Update wormhole adaptor

* add link

* revise and fix test

* deploy

* fix requestResend

* revise

* fix test and re-deploy

* fix test

* fix test for WormholeAdaptor

* revise

---------

+++ contracts/wombat-core/interfaces/IWormhole.sol
@@ -1,4 +1,5 @@
-// Copied from https://github.com/wormhole-foundation/trustless-generic-relayer/blob/2ed875bef7f9113a96cbe4999cfb9d168e6fcd6f/ethereum/contracts/interfaces/IWormhole.sol
+// copied from https://github.com/wormhole-foundation/wormhole/blob/8d63ab50fb7cc80c54fa25b81f764c3a2ee132dc/ethereum/contracts/interfaces/IWormhole.sol
+// contracts/Messages.sol
 // SPDX-License-Identifier: Apache 2
 
 pragma solidity ^0.8.0;
@@ -25,8 +26,10 @@ interface IWormhole {
         uint64 sequence;
         uint8 consistencyLevel;
         bytes payload;
+
         uint32 guardianSetIndex;
         Signature[] signatures;
+
         bytes32 hash;
     }
 
@@ -34,6 +37,7 @@ interface IWormhole {
         bytes32 module;
         uint8 action;
         uint16 chain;
+
         address newContract;
     }
 
@@ -41,6 +45,7 @@ interface IWormhole {
         bytes32 module;
         uint8 action;
         uint16 chain;
+
         GuardianSet newGuardianSet;
         uint32 newGuardianSetIndex;
     }
@@ -49,6 +54,7 @@ interface IWormhole {
         bytes32 module;
         uint8 action;
         uint16 chain;
+
         uint256 messageFee;
     }
 
@@ -56,6 +62,7 @@ interface IWormhole {
         bytes32 module;
         uint8 action;
         uint16 chain;
+
         uint256 amount;
         bytes32 recipient;
     }
@@ -63,38 +70,32 @@ interface IWormhole {
     struct RecoverChainId {
         bytes32 module;
         uint8 action;
+
         uint256 evmChainId;
         uint16 newChainId;
     }
 
-    event LogMessagePublished(
-        address indexed sender, uint64 sequence, uint32 nonce, bytes payload, uint8 consistencyLevel
-    );
+    event LogMessagePublished(address indexed sender, uint64 sequence, uint32 nonce, bytes payload, uint8 consistencyLevel);
     event ContractUpgraded(address indexed oldContract, address indexed newContract);
     event GuardianSetAdded(uint32 indexed index);
 
-    function publishMessage(uint32 nonce, bytes memory payload, uint8 consistencyLevel)
-        external
-        payable
-        returns (uint64 sequence);
+    function publishMessage(
+        uint32 nonce,
+        bytes memory payload,
+        uint8 consistencyLevel
+    ) external payable returns (uint64 sequence);
 
     function initialize() external;
 
-    function parseAndVerifyVM(bytes calldata encodedVM)
-        external
-        view
-        returns (VM memory vm, bool valid, string memory reason);
+    function parseAndVerifyVM(bytes calldata encodedVM) external view returns (VM memory vm, bool valid, string memory reason);
 
     function verifyVM(VM memory vm) external view returns (bool valid, string memory reason);
 
-    function verifySignatures(bytes32 hash, Signature[] memory signatures, GuardianSet memory guardianSet)
-        external
-        pure
-        returns (bool valid, string memory reason);
+    function verifySignatures(bytes32 hash, Signature[] memory signatures, GuardianSet memory guardianSet) external pure returns (bool valid, string memory reason);
 
     function parseVM(bytes memory encodedVM) external pure returns (VM memory vm);
 
-    function quorum(uint256 numGuardians) external pure returns (uint256 numSignaturesRequiredForQuorum);
+    function quorum(uint numGuardians) external pure returns (uint numSignaturesRequiredForQuorum);
 
     function getGuardianSet(uint32 index) external view returns (GuardianSet memory);
 
@@ -122,19 +123,13 @@ interface IWormhole {
 
     function parseContractUpgrade(bytes memory encodedUpgrade) external pure returns (ContractUpgrade memory cu);
 
-    function parseGuardianSetUpgrade(bytes memory encodedUpgrade)
-        external
-        pure
-        returns (GuardianSetUpgrade memory gsu);
+    function parseGuardianSetUpgrade(bytes memory encodedUpgrade) external pure returns (GuardianSetUpgrade memory gsu);
 
     function parseSetMessageFee(bytes memory encodedSetMessageFee) external pure returns (SetMessageFee memory smf);
 
     function parseTransferFees(bytes memory encodedTransferFees) external pure returns (TransferFees memory tf);
 
-    function parseRecoverChainId(bytes memory encodedRecoverChainId)
-        external
-        pure
-        returns (RecoverChainId memory rci);
+    function parseRecoverChainId(bytes memory encodedRecoverChainId) external pure returns (RecoverChainId memory rci);
 
     function submitContractUpgrade(bytes memory _vm) external;
